[MSTDungeon] Minor improvements to bring back visuals
Previous iterations had visual feedback which was lost later because of
refactoring.

I'm bringing them back, starting with real-time TileMap animation based
on the physics simulation.

Next parts will be with visual feedback on the graph, creating the MST
graph, pruning it and keeping certian rooms.

Changes in this commit:

- add TileMap real-time visual feedback during physics simulation
- add iterator functions to Room class