Update blue noise to use grids that fit asteroid counts closer
We would use a grid the square of the number of asteroids to generate to
randomly place asteroids. With high densities, we would lose the
blue-ish noise quality (uniform distribution).

This change makes the code use the closest upper square that fits all
asteroids.

E.g. for 3 asteroids, the grid will be 2x2, and for 6 asteroids, it'll
be 3x3